Query order and role columns

+++ internal/role/role.go
@@ -234,15 +234,18 @@ func (r *Role) Create(super bool) (out []postgres.SyncQuery) {
 
 func (r *Role) Drop(databases *postgres.DBMap, currentUser Role, fallbackOwner string) (out []postgres.SyncQuery) {
 	identifier := pgx.Identifier{r.Name}
-	out = append(out, postgres.SyncQuery{
-		Description: "Terminate running sessions.",
-		LogArgs:     []interface{}{"role", r.Name},
-		Query: `
-		SELECT pg_terminate_backend(pid)
-		FROM pg_catalog.pg_stat_activity
-		WHERE usename = %s;`,
-		QueryArgs: []interface{}{r.Name},
-	})
+	if r.Options.CanLogin {
+		out = append(out, postgres.SyncQuery{
+			Description: "Terminate running sessions.",
+			LogArgs:     []interface{}{"role", r.Name},
+			Database:    "<first>",
+			Query: `
+			SELECT pg_terminate_backend(pid)
+			FROM pg_catalog.pg_stat_activity
+			WHERE usename = %s;`,
+			QueryArgs: []interface{}{r.Name},
+		})
+	}
 
 	if !currentUser.Options.Super {
 		// Non-super user needs to inherit to-be-dropped role to reassign objects.
@@ -253,7 +256,8 @@ func (r *Role) Drop(databases *postgres.DBMap, currentUser Role, fallbackOwner s
 				LogArgs: []interface{}{
 					"role", r.Name, "parent", currentUser.Name,
 				},
-				Query: `REVOKE %s FROM %s;`,
+				Database: "<first>",
+				Query:    `REVOKE %s FROM %s;`,
 				QueryArgs: []interface{}{
 					pgx.Identifier{currentUser.Name},
 					identifier,
@@ -265,7 +269,8 @@ func (r *Role) Drop(databases *postgres.DBMap, currentUser Role, fallbackOwner s
 			LogArgs: []interface{}{
 				"role", r.Name, "parent", currentUser.Name,
 			},
-			Query: `GRANT %s TO %s;`,
+			Database: "<first>",
+			Query:    `GRANT %s TO %s;`,
 			QueryArgs: []interface{}{
 				identifier,
 				pgx.Identifier{currentUser.Name},
@@ -277,9 +282,9 @@ func (r *Role) Drop(databases *postgres.DBMap, currentUser Role, fallbackOwner s
 			out = append(out, postgres.SyncQuery{
 				Description: "Reassign database.",
 				LogArgs: []interface{}{
-					"role", r.Name,
 					"database", database.Name,
-					"owner", fallbackOwner,
+					"old", r.Name,
+					"new", fallbackOwner,
 				},
 				Query: `ALTER DATABASE %s OWNER TO %s;`,
 				QueryArgs: []interface{}{
